requirements.txt
boto3==1.38.15
botocore==1.38.15
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
logging==0.4.9.6
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pluggy==1.5.0
pytest==8.3.5
pytest-assume==2.4.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
